# Adds perf tests for `Map.merge`.