--- a/perf/Map.js
+++ b/perf/Map.js
@@ -141,24 +141,4 @@ describe('Map', function() {
       });
     });
   });
-
-  describe('mergeDeep a map', () => {
-    [2, 8, 32, 1024].forEach(size => {
-      const obj1 = {};
-      const obj2 = {};
-      for (let ii = 0; ii < size; ii++) {
-        obj1['k' + ii] = {};
-        obj1['k' + ii]['k' + ii] = '1_' + ii;
-        obj2['k' + ii] = {};
-        obj2['k' + ii]['k' + ii] = '2_' + ii;
-      }
-
-      const map1 = Immutable.Map(obj1);
-      const map2 = Immutable.Map(obj2);
-
-      it('of ' + size, () => {
-        map1.merge(map2);
-      });
-    });
-  });
 });