Output sample:
```
Map > merge an object of 2
  Old:   932,047   960,634   991,030 ops/sec
  New:   978,449   983,047   987,688 ops/sec
  compare: -1 1
  diff: 2.33%
  rme: 2.19%
Map > merge an object of 8
  Old:   568,422   572,977   577,606 ops/sec
  New:   571,550   575,883   580,283 ops/sec
  compare: 0 0
  diff: 0.5%
  rme: 0.78%
Map > merge an object of 32
  Old:   251,965   254,599   257,289 ops/sec
  New:   228,904   232,267   235,731 ops/sec
  compare: 1 -1
  diff: -8.78%
  rme: 1.27%
Map > merge an object of 1024
  Old:     8,428     8,491     8,554 ops/sec
  New:     7,625     7,681     7,739 ops/sec
  compare: 1 -1
  diff: -9.54%
  rme: 0.73%
Map > mergeDeep an object of 2
  Old:   321,103   323,956   326,860 ops/sec
  New:   323,939   328,831   333,872 ops/sec
  compare: -1 1
  diff: 1.5%
  rme: 1.23%
Map > mergeDeep an object of 8
  Old:   125,157   126,434   127,736 ops/sec
  New:   124,574   126,283   128,040 ops/sec
  compare: 0 0
  diff: -0.12%
  rme: 1.2%
Map > mergeDeep an object of 32
  Old:    37,218    37,439    37,662 ops/sec
  New:    36,882    37,395    37,923 ops/sec
  compare: 0 0
  diff: -0.12%
  rme: 1.06%
Map > mergeDeep an object of 1024
  Old:       728       740       752 ops/sec
  New:       698       712       726 ops/sec
  compare: 1 -1
  diff: -3.82%
  rme: 1.8%
```

Full output: [map-merge-perf.txt](https://github.com/facebook/immutable-js/files/1395673/map-merge-perf.txt)