@abacaphiliac 2017-10-18
@leebyron Done. Thanks!

Are you interested in me getting the perf tests running in Travis?