@leebyron 2017-10-18
I think the perf tests might be too slow to run in Travis, also they're pretty sensitive the the operating environment so I think they may also be too noisy