@leebyron 2017-10-18
Thanks! I added a few additional changes. I followed up with the change I mentioned before about ensuring the second map contains different values. I also removed `mergeDeep()` since it seemed broken (it wasn't calling `mergeDeep()`) and that's just a thin layer atop `merge()` anyhow, so from a perf testing point of view, it's probably duplicative